requirements.txt
streamlit
os
shutil
time
langchain
langchain-cohere
trulens-eval
textblob
PyPDF2
chroma
ollama
langchain-community
langchain-core
langchain-google-genai
python-dotenv
